Type job salaries as numeric, not integer
careerhub/schema.sql declares these integer and the crawler's NormalizedJob
types them Optional[int], so 023a followed suit — but career-hub's live table
had drifted to numeric, and 238 of its salary values carry a non-zero fraction.
Loading them into an integer column fails outright, and rounding to fit would
silently discard hourly rates like 13.50.

numeric accepts everything integer does, so the crawler writing ints is
unaffected.

+++ supabase/migrations/023a_careerhub_base.sql
@@ -18,8 +18,12 @@ create table if not exists public.jobs (
   title         text,
   location_raw  text,
   city          text,
-  salary_min    integer,
-  salary_max    integer,
+  -- `careerhub/schema.sql` says integer and the crawler's NormalizedJob types
+  -- these as Optional[int], but career-hub's live table had drifted to numeric
+  -- and holds fractional hourly rates — 238 of them. integer would round those
+  -- away on any future import, and numeric costs nothing to accept ints.
+  salary_min    numeric,
+  salary_max    numeric,
   posted_at     timestamptz,
   is_active     boolean default true,
   source        text,
